+++ mycli/client_connection.py
@@ -31,7 +31,7 @@ except ImportError:
 
 class ClientConnectionMixin:
     if TYPE_CHECKING:
-        my_cnf: Any
+        mylogin_cnf: Any
         config: Any
         config_without_package_defaults: Any
         keepalive_ticks: int | None
@@ -39,8 +39,7 @@ class ClientConnectionMixin:
         sqlexecute: Any
         logger: Any
 
-        def read_my_cnf(self, files: Any, keys: list[str]) -> dict[str, Any]: ...
-        def merge_ssl_with_cnf(self, ssl_config: dict[str, Any], cnf: dict[str, Any]) -> dict[str, Any] | None: ...
+        def read_mylogin_cnf(self, cnf: Any) -> dict[str, Any]: ...
         def echo(self, *args: Any, **kwargs: Any) -> None: ...
 
     def connect(
@@ -60,31 +59,11 @@ class ClientConnectionMixin:
         reset_keyring: bool | None = None,
         keepalive_ticks: int | None = None,
     ) -> None:
-        cnf = {
-            "database": None,
-            "user": None,
-            "password": None,
-            "host": None,
-            "port": None,
-            "socket": None,
-            "default_socket": None,
-            "default-character-set": None,
-            "local-infile": None,
-            "loose-local-infile": None,
-            "ssl-ca": None,
-            "ssl-cert": None,
-            "ssl-key": None,
-            "ssl-cipher": None,
-            "ssl-verify-server-cert": None,
-        }
-
-        cnf = self.read_my_cnf(self.my_cnf, list(cnf.keys()))
-
-        # Fall back to config values only if user did not specify a value.
-        database = database or cnf["database"]
-        user = user or cnf["user"] or os.getenv("USER")
-        host = host or cnf["host"]
-        port = port or cnf["port"]
+        mylogin_cnf: dict[str, Any] = self.read_mylogin_cnf(self.mylogin_cnf)
+        # Fall back to .mylogin.cnf values only if user did not specify a value.
+        user = user or mylogin_cnf["user"] or os.getenv("USER")
+        host = host or mylogin_cnf["host"]
+        port = port or mylogin_cnf["port"]
         ssl_config: dict[str, Any] = ssl or {}
         user_connection_config = self.config_without_package_defaults.get('connection', {})
         self.keepalive_ticks = keepalive_ticks
@@ -93,15 +72,9 @@ class ClientConnectionMixin:
         if not int_port:
             int_port = DEFAULT_PORT
             if not host or host == DEFAULT_HOST:
-                socket = (
-                    socket
-                    or user_connection_config.get("default_socket")
-                    or cnf["socket"]
-                    or cnf["default_socket"]
-                    or guess_socket_location()
-                )
+                socket = socket or user_connection_config.get("default_socket") or mylogin_cnf["socket"] or guess_socket_location()
 
-        passwd = passwd if isinstance(passwd, (str, int)) else cnf["password"]
+        passwd = passwd if isinstance(passwd, (str, int)) else mylogin_cnf["password"]
 
         # default_character_set doesn't check in self.config_without_package_defaults, because the
         # option already existed before the my.cnf deprecation.  For the same reason,
@@ -111,10 +84,6 @@ class ClientConnectionMixin:
                 character_set = self.config['connection']['default_character_set']
             elif 'default_character_set' in self.config['main']:
                 character_set = self.config['main']['default_character_set']
-            elif 'default_character_set' in cnf:
-                character_set = cnf['default_character_set']
-            elif 'default-character-set' in cnf:
-                character_set = cnf['default-character-set']
         if not character_set:
             character_set = DEFAULT_CHARSET
 
@@ -123,10 +92,6 @@ class ClientConnectionMixin:
         for local_infile_option in (
             local_infile,
             user_connection_config.get('default_local_infile'),
-            cnf['local_infile'],
-            cnf['local-infile'],
-            cnf['loose_local_infile'],
-            cnf['loose-local-infile'],
             False,
         ):
             try:
@@ -135,37 +100,19 @@ class ClientConnectionMixin:
             except (TypeError, ValueError):
                 pass
 
-        # temporary my.cnf override mappings
-        if 'default_ssl_ca' in user_connection_config:
-            cnf['ssl-ca'] = user_connection_config.get('default_ssl_ca') or None
-        if 'default_ssl_cert' in user_connection_config:
-            cnf['ssl-cert'] = user_connection_config.get('default_ssl_cert') or None
-        if 'default_ssl_key' in user_connection_config:
-            cnf['ssl-key'] = user_connection_config.get('default_ssl_key') or None
-        if 'default_ssl_cipher' in user_connection_config:
-            cnf['ssl-cipher'] = user_connection_config.get('default_ssl_cipher') or None
-        if 'default_ssl_verify_server_cert' in user_connection_config:
-            cnf['ssl-verify-server-cert'] = user_connection_config.get('default_ssl_verify_server_cert') or None
-
-        # todo: rewrite the merge method using self.config['connection'] instead of cnf, after removing my.cnf support
-        ssl_config_or_none: dict[str, Any] | None = self.merge_ssl_with_cnf(ssl_config, cnf)
-
-        # default_ssl_ca_path is not represented in my.cnf
-        if 'default_ssl_ca_path' in self.config['connection'] and (not ssl_config_or_none or not ssl_config_or_none.get('capath')):
-            if ssl_config_or_none is None:
-                ssl_config_or_none = {}
-            ssl_config_or_none['capath'] = self.config['connection']['default_ssl_ca_path'] or False
+        if 'default_ssl_ca_path' in self.config['connection'] and (not ssl_config or not ssl_config.get('capath')):
+            ssl_config['capath'] = self.config['connection']['default_ssl_ca_path'] or False
 
         # prune lone check_hostname=False
         if not any(v for v in ssl_config.values()):
-            ssl_config_or_none = None
+            ssl_config = {}
 
         # password hierarchy
         # 1. -p / --pass/--password CLI options
         # 2. --password-file CLI option
         # 3. envvar (MYSQL_PWD)
         # 4. DSN (mysql://user:password)
-        # 5. cnf (.my.cnf / etc)
+        # 5. .mylogin.cnf
         # 6. keyring
 
         keyring_identifier = f'{user}@{host}:{"" if socket else int_port}:{socket or ""}'
@@ -194,7 +141,7 @@ class ClientConnectionMixin:
             "socket": socket,
             "character_set": character_set,
             "local_infile": use_local_infile,
-            "ssl": ssl_config_or_none,
+            "ssl": ssl_config,
             "init_command": init_command,
             "unbuffered": unbuffered,
         }